+++ gui/tooltip-reader.lua
@@ -1,2 +1,0 @@
-EasyBronze.tooltipReader = CreateFrame('GameTooltip', "EasyBronzeTooltipReader", nil, "GameTooltipTemplate")
-EasyBronze.tooltipReader:SetOwner(UIParent, "ANCHOR_NONE")