+++ gui/tooltip-scanner.lua
@@ -1,0 +1,2 @@
+EasyBronze.tooltipScanner = CreateFrame('GameTooltip', "EasyBronzeTooltipScanner", nil, "GameTooltipTemplate")
+EasyBronze.tooltipScanner:SetOwner(UIParent, "ANCHOR_NONE")